requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.20
botocore==1.31.20
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
cloudinary==1.34.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
google-auth==2.23.3
gunicorn==21.2.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.0
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
redis-cli==1.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
service-identity==23.1.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
stripe==6.2.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
zope.interface==6.1
